fix(GitHub): Create release with change logs with specific version tag
