+++ .github/scripts/changelogs/basic_commit_parser.py
@@ -1,0 +1,53 @@
+
+import logging
+import re
+from typing import Dict, Optional
+
+from base_interfaces import CommitParser
+
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
+
+class BasicCommitParser(CommitParser):
+    TYPES = ["feat", "fix", "docs", "style", "refactor", "perf", "test", "chore"]
+
+    def __init__(self):
+        self.commit_pattern = self._create_commit_pattern()
+
+    def _create_commit_pattern(self):
+        type_regex = "|".join(self.TYPES)
+        return re.compile(
+            r"^(?P<type>" + type_regex + r")"
+            r"\((?P<scope>[^)]+)\):\s*"
+            r"(?P<title>[^\n]+)"
+            r"(?:(?P<body>[\s\S]*?))?"
+            r"(?:\nRefs:\s*(?P<refs>#[A-Za-z0-9-]+(?:,\s*#[A-Za-z0-9-]+)*))?"
+            r"$", re.DOTALL
+        )
+
+    def parse(self, message: str) -> Optional[Dict]:
+        if not message or not isinstance(message, str):
+            logger.warning(f"Invalid message format: {message}")
+            return None
+
+        message = message.strip()
+        match = self.commit_pattern.match(message)
+
+        if not match:
+            logger.debug(f"No match found for message: {message}")
+            return None
+
+        result = match.groupdict()
+
+        return {
+            "type": result.get("type", ""),
+            "scope": result.get("scope", ""),
+            "title": result.get("title", ""),
+            "body": result.get("body", "").strip() if result.get("body") else None,
+            "refs": [ref.strip() for ref in result.get("refs", "").split(",")] if result.get("refs") else []
+        }
+
