+++ .github/scripts/changelogs/commit_fetcher.py
@@ -1,0 +1,48 @@
+import requests
+
+from base_interfaces import CommitFetcher
+from typing import Dict, List
+import logging
+
+
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
+
+class GitHubCommitFetcher(CommitFetcher):
+    def __init__(self, github_token: str, repo_owner: str, repo_name: str):
+        self.github_token = github_token
+        self.repo_owner = repo_owner
+        self.repo_name = repo_name
+
+    def fetch_commits(self, branch="main") -> List[Dict]:
+        url = f"https://api.github.com/repos/{self.repo_owner}/{self.repo_name}/commits"
+        headers = {
+            "Authorization": f"token {self.github_token}",
+            "Accept": "application/vnd.github.v3+json"
+        }
+
+        commits = []
+        page = 1
+
+        while True:
+            try:
+                response = requests.get(
+                    url,
+                    headers=headers,
+                    params={"sha": branch, "page": page}
+                )
+                response.raise_for_status()
+                data = response.json()
+                if not data:
+                    break
+                commits.extend(data)
+                page += 1
+            except Exception as e:
+                logger.error(f"Error fetching commits: {e}")
+                break
+
+        return commits
