+++ .github/scripts/changelogs/enhanced_commit_document_manager.py
@@ -1,0 +1,41 @@
+import os
+
+from base_interfaces import CommitDocumentManager
+from commit_fetcher import CommitFetcher
+from .basic_commit_parser import CommitParser
+from report_generator_factory import ReportGeneratorFactory
+from datetime import datetime
+import logging
+
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
+
+class EnhancedCommitDocumentManager(CommitDocumentManager):
+    def __init__(self, commit_fetcher: CommitFetcher, commit_parser: CommitParser):
+        super().__init__(commit_fetcher, commit_parser)
+        self.reports = {}
+
+    def generate_all_documents(self):
+        commits = self.commit_fetcher.fetch_commits()
+        categorized = self.categorize_commits(commits)
+        
+        generators = {
+            'release_notes': ReportGeneratorFactory.create_generator('release'),
+            'commit_report': ReportGeneratorFactory.create_generator('markdown')
+        }
+
+        for report_name, generator in generators.items():
+            content = generator.generate(categorized)
+            self.save_document(content, f"generated_docs/{report_name}.md")
+
+    def save_document(self, content: str, base_filename: str):
+        filename = f"{os.path.splitext(base_filename)[0]}_{datetime.now().strftime('%Y-%m-%d')}.md"
+        os.makedirs(os.path.dirname(filename), exist_ok=True)
+        
+        with open(filename, 'w', encoding='utf-8') as file:
+            file.write(content)
+        logger.info(f"✅ Generated {filename}")
