+++ .github/scripts/changelogs/main.py
@@ -1,0 +1,41 @@
+import os
+import logging
+
+from commit_fetcher import GitHubCommitFetcher
+from .basic_commit_parser import BasicCommitParser
+from enhanced_commit_document_manager import EnhancedCommitDocumentManager
+
+
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
+
+
+def main():
+    # Setup GitHub credentials
+    github_token = os.getenv('GITHUB_TOKEN')
+    repo_owner = os.getenv('REPO_OWNER')
+    repo_name = os.getenv('REPO_NAME')
+
+    if not all([github_token, repo_owner, repo_name]):
+        logger.error("Missing required environment variables")
+        return
+
+    try:
+        # Create components
+        commit_fetcher = GitHubCommitFetcher(github_token, repo_owner, repo_name)
+        commit_parser = BasicCommitParser()
+
+        # Create document manager and generate reports
+        document_manager = EnhancedCommitDocumentManager(commit_fetcher, commit_parser)
+        document_manager.generate_all_documents()
+        
+        logger.info("✅ Successfully generated all reports")
+    except Exception as e:
+        logger.error(f"Error generating reports: {e}")
+
+if __name__ == "__main__":
+    main()
