+++ .github/workflows/generate-doc.yml
@@ -30,7 +30,7 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           REPO_OWNER: ${{ github.repository_owner }}  # Set owner as environment variable
           REPO_NAME: ${{ github.event.repository.name }}  # Set repo name as environment variable
-        run: python .github/scripts/generate_commit_doc.py
+        run: python .github/scripts/changelogs/main.py
 
       - name: Check the generated commit logs
         run: ls -la generated_docs  # This is to check that files exist in the folder
